+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project follows [Semantic Versioning](http://semver.org) guidelines.
 ### Fixed
 * Fixes an exception thrown when trying to access the `/nests/:id/eggs/:id` API endpoint.
 
+### Added
+* Adds ability to include egg variables on an API request.
+
 ## v0.7.1 (Derelict Dermodactylus)
 ### Fixed
 * Fixes an exception when no token is entered on the 2-Factor enable/disable page and the form is submitted.